Introduce account search.py file
