+++ saleor/account/utils.py
@@ -84,22 +84,3 @@ def remove_staff_member(staff):
         staff.save()
     else:
         staff.delete()
-
-
-def prepare_user_search_document_value(user: "User", *, attach_addresses_data=True):
-    """Prepare `search_document` user value - attach all field used in searching.
-
-    Parameter `attach_addresses_data` should be set to False only when user
-    is created and no address has been attached or user addresses are cleared.
-    """
-    search_document = f"{user.email}{user.first_name}{user.last_name}"
-
-    if attach_addresses_data:
-        for address in user.addresses.all():
-            search_document += (
-                f"{address.first_name}{address.last_name}"
-                f"{address.street_address_1}{address.street_address_2}"
-                f"{address.city}{address.postal_code}{address.country}{address.phone}"
-            )
-
-    return search_document.replace(" ", "").lower()
