+++ saleor/graphql/account/mutations/staff.py
@@ -8,8 +8,9 @@ from ....account import events as account_events
 from ....account import models, utils
 from ....account.error_codes import AccountErrorCode
 from ....account.notifications import send_set_password_notification
+from ....account.search import USER_SEARCH_FIELDS, prepare_user_search_document_value
 from ....account.thumbnails import create_user_avatar_thumbnails
-from ....account.utils import prepare_user_search_document_value, remove_staff_member
+from ....account.utils import remove_staff_member
 from ....checkout import AddressType
 from ....core.exceptions import PermissionDenied
 from ....core.permissions import AccountPermissions
@@ -245,9 +246,7 @@ class StaffCreate(ModelMutation):
 
     @classmethod
     def save(cls, info, user, cleaned_input):
-        if any(
-            [field in cleaned_input for field in ["email", "first_name", "last_name"]]
-        ):
+        if any([field in cleaned_input for field in USER_SEARCH_FIELDS]):
             user.search_document = prepare_user_search_document_value(
                 user, attach_addresses_data=False
             )
